+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Date format: DD/MM/YYYY
 - Fix: [#207](https://github.com/bdlukaa/fluent_ui/pull/207) FilledButton disabled foreground
 - NEW: Widgets translated into multiple languages: [#216](https://github.com/bdlukaa/fluent_ui/pull/216)
 - Add `useInheritedMediaQuery` property to `FluentApp` [#211](https://github.com/bdlukaa/fluent_ui/pull/211)
+- TreeView now implemented with ListView and optionally allows vertical scrolling ([#255](https://github.com/bdlukaa/fluent_ui/pull/225))
+- TreeViewItem now has custom primary key (`value` field)
+- Added `onSelectionChanged` callback to `TreeView`
 
 ## [3.9.1] - Input Update - [25/02/2022]